backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.8
django-celery-beat==2.6.0
whitenoise==6.7.0
gunicorn==22.0.0
python-dotenv==1.0.1
